implemented add note functionality

+++ app/src/main/java/com/takeNote/feature_note/data/data_source/NoteDao.kt
@@ -11,7 +11,7 @@ interface NoteDao {
     fun getNotes(): Flow<List<Note>>
 
     @Query("SELECT * FROM note WHERE id = :id")
-    suspend fun getNoteById(id: Long): Note?
+    suspend fun getNoteById(id: Int): Note?
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     suspend fun insertNote(note: Note)
