+++ app/src/main/java/com/takeNote/feature_note/domain/repository/NoteRepository.kt
@@ -5,7 +5,7 @@ import kotlinx.coroutines.flow.Flow
 
 interface NoteRepository {
     fun getAllNotes(): Flow<List<Note>>
-    suspend fun getNoteById(id: Long): Note?
+    suspend fun getNoteById(id: Int): Note?
     suspend fun insertNote(note: Note)
     suspend fun deleteNote(note: Note)
 }
